fix: 전체 모듈 리뷰 발견사항 수정 #42
자산 환전 로그 입금액을 수수료 차감 후 최종 증가액으로 저장하고 기존 로그 보정 업데이트를 추가했다.

콘텐츠 다운로드와 커뮤니티 첨부는 파일 전달 준비가 끝난 뒤에만 자산을 차감하도록 순서를 조정했다.

쿠폰 관리자 입력의 key, 중복 key, 사용 가능 횟수 서버 검증을 보강하고 회귀 방지 점검과 리뷰 기록을 추가했다.

+++ modules/coupon/actions/admin-coupons.php
@@ -29,7 +29,7 @@ if (sr_request_method() === 'POST') {
                 'target_type' => sr_post_string('target_type', 60),
                 'target_id' => sr_post_string('target_id', 80),
                 'refundable_policy' => sr_post_string('refundable_policy', 30),
-                'max_uses_per_issue' => (int) sr_post_string('max_uses_per_issue', 10),
+                'max_uses_per_issue' => sr_post_string('max_uses_per_issue', 10),
             ]);
             sr_audit_log($pdo, [
                 'actor_account_id' => (int) $account['id'],
